Move DailyServiceTimes types to lib-common
